resource: isolate fractional opt
